tests/wpt/tests/html/semantics/scripting-1/the-script-element/nomodule-set-on-inline-module-script.html

<!DOCTYPE html>
<html>
<head>
<title>An inline module script with nomodule must run</title>
<link rel="author" title="Yusuke Suzuki" href="mailto:utatane.tea@gmail.com">
<link rel="author" title="Ryosuke Niwa" href="mailto:rniwa@webkit.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script nomodule type="module">
import Cocoa from "./resources/cocoa-module.js";
var cocoa = new Cocoa();
window.exportedCocoa = cocoa;
</script>
<script>

waitForLoadEvent = new Promise((resolve) => {
    window.onload = resolve;
});

promise_test(() => {
    return waitForLoadEvent.then(() => {
        assert_equals(typeof cocoa, 'undefined');
        assert_equals(typeof exportedCocoa, 'object');
        assert_equals(exportedCocoa.taste(), 'awesome');
    });
}, 'An inline module script with nomodule content attribute must run');

</script>
</body>
</html>
