third_party/WebKit/LayoutTests/external/wpt/html/semantics/scripting-1/the-script-element/nomodule-set-on-inline-module-script-expected.txt

CONSOLE ERROR: line 12: Uncaught SyntaxError: Unexpected token import
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught SyntaxError: Unexpected token import
FAIL An inline module script with nomodule content attribute must run assert_equals: expected "object" but got "undefined"
Harness: the test ran to completion.
